fixes from spec changes

+++ openapi/models/operations/listdetails.go
@@ -51,7 +51,7 @@ type ListDetails200ApplicationJSONItems struct {
 	ReleaseDate      *string  `json:"release_date,omitempty"`
 	Title            *string  `json:"title,omitempty"`
 	Video            *bool    `default:"true" json:"video"`
-	VoteAverage      *int64   `default:"0" json:"vote_average"`
+	VoteAverage      *float64 `default:"0" json:"vote_average"`
 	VoteCount        *int64   `default:"0" json:"vote_count"`
 }
 
@@ -157,7 +157,7 @@ func (o *ListDetails200ApplicationJSONItems) GetVideo() *bool {
 	return o.Video
 }
 
-func (o *ListDetails200ApplicationJSONItems) GetVoteAverage() *int64 {
+func (o *ListDetails200ApplicationJSONItems) GetVoteAverage() *float64 {
 	if o == nil {
 		return nil
 	}
