+++ openapi/models/operations/tvseriesairingtodaylist.go
@@ -57,7 +57,7 @@ type TvSeriesAiringTodayList200ApplicationJSONResults struct {
 	Overview         *string  `json:"overview,omitempty"`
 	Popularity       *float64 `default:"0" json:"popularity"`
 	PosterPath       *string  `json:"poster_path,omitempty"`
-	VoteAverage      *int64   `default:"0" json:"vote_average"`
+	VoteAverage      *float64 `default:"0" json:"vote_average"`
 	VoteCount        *int64   `default:"0" json:"vote_count"`
 }
 
@@ -149,7 +149,7 @@ func (o *TvSeriesAiringTodayList200ApplicationJSONResults) GetPosterPath() *stri
 	return o.PosterPath
 }
 
-func (o *TvSeriesAiringTodayList200ApplicationJSONResults) GetVoteAverage() *int64 {
+func (o *TvSeriesAiringTodayList200ApplicationJSONResults) GetVoteAverage() *float64 {
 	if o == nil {
 		return nil
 	}
