+++ openapi/models/operations/tvseriessimilar.go
@@ -58,7 +58,7 @@ type TvSeriesSimilar200ApplicationJSONResults struct {
 	Overview         *string  `json:"overview,omitempty"`
 	Popularity       *float64 `default:"0" json:"popularity"`
 	PosterPath       *string  `json:"poster_path,omitempty"`
-	VoteAverage      *int64   `default:"0" json:"vote_average"`
+	VoteAverage      *float64 `default:"0" json:"vote_average"`
 	VoteCount        *int64   `default:"0" json:"vote_count"`
 }
 
@@ -157,7 +157,7 @@ func (o *TvSeriesSimilar200ApplicationJSONResults) GetPosterPath() *string {
 	return o.PosterPath
 }
 
-func (o *TvSeriesSimilar200ApplicationJSONResults) GetVoteAverage() *int64 {
+func (o *TvSeriesSimilar200ApplicationJSONResults) GetVoteAverage() *float64 {
 	if o == nil {
 		return nil
 	}
